+++ StudentManagerment2/Controllers/StudentController.cs
@@ -1,29 +1,32 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
+﻿using System.Linq;
 using System.Web.Mvc;
 using StudentManagerment2.Models;
 
 namespace StudentManagerment2.Controllers
 {
-    
+
     public class StudentController : Controller
     {
-        // GET: Student
-        private ApplicationDbContext db = new ApplicationDbContext();
+        private readonly ApplicationDbContext db = new ApplicationDbContext();//kết nối với database
+
+        
+
         public ActionResult Index()
         {
-            var students = db.Students.Include("Class").ToList();
+            var students = db.Students.Include("Class").ToList();//lấy ra danh sách sinh viên
             return View(students);
         }
-        [AuthorizeByRole("Admin")]//test
+
+
         
+        [AuthorizeByRole("Admin")]
         public ActionResult Create()
         {
-            ViewBag.ClassId = new SelectList(db.Classes, "Id", "ClassName");
+            ViewBag.ClassId = new SelectList(db.Classes, "Id", "ClassName");//tạo ra danh sách lớp
             return View();
         }
+
+        
         [HttpPost]
         [AuthorizeByRole("Admin")]
         [ValidateAntiForgeryToken]
@@ -35,54 +38,74 @@ namespace StudentManagerment2.Controllers
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            ViewBag.ClassId = new SelectList(db.Classes, "Id", "ClassName", student.ClassId);
+            ViewBag.ClassId = new SelectList(db.Classes, "Id", "ClassName", student.ClassId);//tạo ra danh sách lớp
             return View(student);
         }
-        [AuthorizeByRole("Admin")]
-        public ActionResult Delete(int id)
-        {
-            var student = db.Students.Find(id);
-            return View(student);
-        }
-        [HttpPost]
-        [AuthorizeByRole("Admin")]
-        public ActionResult Delete(int id, Student student)
-        {
-            student = db.Students.Find(id);
-            db.Students.Remove(student);
-            db.SaveChanges();
-            return RedirectToAction("Index");
-        }
+
+        
         [AuthorizeByRole("Admin")]
         public ActionResult Edit(int? id)
         {
             if (id == null)
             {
-                return new HttpStatusCodeResult(System.Net.HttpStatusCode.BadRequest);
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.BadRequest);//trả về lỗi 400
             }
-            Student student = db.Students.Find(id);
+            var student = db.Students.Find(id);
             if (student == null)
             {
                 return HttpNotFound();
             }
-            ViewBag.ClassId = new SelectList(db.Classes, "Id", "ClassName", student.ClassId);
+            ViewBag.ClassId = new SelectList(db.Classes, "Id", "ClassName", student.ClassId);//tạo ra danh sách lớp
             return View(student);
         }
 
+        
         [HttpPost]
         [AuthorizeByRole("Admin")]
-        [ValidateAntiForgeryToken]
+        [ValidateAntiForgeryToken]// chống tấn công CSRF
         public ActionResult Edit(Student student)
         {
             if (ModelState.IsValid)
             {
-                db.Entry(student).State = System.Data.Entity.EntityState.Modified;
+                db.Entry(student).State = System.Data.Entity.EntityState.Modified;//cập nhật dữ liệu
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
-            ViewBag.ClassId = new SelectList(db.Classes, "Id", "ClassName", student.ClassId);
+            ViewBag.ClassId = new SelectList(db.Classes, "Id", "ClassName", student.ClassId);//tạo ra danh sách lớp
             return View(student);
         }
 
+        
+        [AuthorizeByRole("Admin")]
+        public ActionResult Delete(int? id)
+        {
+            if (id == null)
+            {
+                return new HttpStatusCodeResult(System.Net.HttpStatusCode.BadRequest);
+            }
+            var student = db.Students.Find(id);
+            if (student == null)
+            {
+                return HttpNotFound();
+            }
+
+            return View(student);
+        }
+
+        
+        [HttpPost, ActionName("Delete")]
+        [AuthorizeByRole("Admin")]
+        [ValidateAntiForgeryToken]
+        public ActionResult DeleteConfirmed(int id)
+        {
+            var student = db.Students.Find(id);
+            if (student != null)
+            {
+                db.Students.Remove(student);
+                db.SaveChanges();
+            }
+            return RedirectToAction("Index");
+        }
     }
 }
+
