+++ StudentManagerment2/Controllers/SubjectController.cs
@@ -4,26 +4,26 @@ using StudentManagerment2.Models;
 
 namespace StudentManagerment2.Controllers
 {
-    
+
     public class SubjectController : Controller
     {
-        private ApplicationDbContext db = new ApplicationDbContext();
+        private readonly ApplicationDbContext db = new ApplicationDbContext();//khai báo db context
 
-        // GET: Subject
+        
         public ActionResult Index()
         {
             var subjects = db.Subjects.ToList();
             return View(subjects);
         }
 
-        // GET: Subject/Create
+        
         [AuthorizeByRole("Admin")]
         public ActionResult Create()
         {
             return View();
         }
 
-        // POST: Subject/Create
+        
         [HttpPost]
         [AuthorizeByRole("Admin")]
         [ValidateAntiForgeryToken]
@@ -38,7 +38,7 @@ namespace StudentManagerment2.Controllers
             return View(subject);
         }
 
-        // GET: Subject/Edit/5
+        
         [AuthorizeByRole("Admin")]
         public ActionResult Edit(int? id)
         {
@@ -46,7 +46,7 @@ namespace StudentManagerment2.Controllers
             {
                 return new HttpStatusCodeResult(System.Net.HttpStatusCode.BadRequest);
             }
-            Subject subject = db.Subjects.Find(id);
+            var subject = db.Subjects.Find(id);
             if (subject == null)
             {
                 return HttpNotFound();
@@ -54,7 +54,7 @@ namespace StudentManagerment2.Controllers
             return View(subject);
         }
 
-        // POST: Subject/Edit/5
+        
         [HttpPost]
         [AuthorizeByRole("Admin")]
         [ValidateAntiForgeryToken]
@@ -62,14 +62,14 @@ namespace StudentManagerment2.Controllers
         {
             if (ModelState.IsValid)
             {
-                db.Entry(subject).State = System.Data.Entity.EntityState.Modified;
+                db.Entry(subject).State = System.Data.Entity.EntityState.Modified;//cập nhật dữ liệu
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
             return View(subject);
         }
 
-        // GET: Subject/Delete/5
+        
         [AuthorizeByRole("Admin")]
         public ActionResult Delete(int? id)
         {
@@ -77,7 +77,7 @@ namespace StudentManagerment2.Controllers
             {
                 return new HttpStatusCodeResult(System.Net.HttpStatusCode.BadRequest);
             }
-            Subject subject = db.Subjects.Find(id);
+            var subject = db.Subjects.Find(id);
             if (subject == null)
             {
                 return HttpNotFound();
@@ -85,16 +85,20 @@ namespace StudentManagerment2.Controllers
             return View(subject);
         }
 
-        // POST: Subject/Delete/5
+        
         [HttpPost, ActionName("Delete")]
         [AuthorizeByRole("Admin")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Subject subject = db.Subjects.Find(id);
-            db.Subjects.Remove(subject);
-            db.SaveChanges();
+            var subject = db.Subjects.Find(id);
+            if (subject != null)
+            {
+                db.Subjects.Remove(subject);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
     }
 }
+
