+++ StudentManagerment2/Controllers/TeacherController.cs
@@ -4,26 +4,26 @@ using StudentManagerment2.Models;
 
 namespace StudentManagerment2.Controllers
 {
-    
+
     public class TeacherController : Controller
     {
-        private ApplicationDbContext db = new ApplicationDbContext();
+        private readonly ApplicationDbContext db = new ApplicationDbContext();
 
-        // GET: Teacher
+        
         public ActionResult Index()
         {
             var teachers = db.Teachers.ToList();
             return View(teachers);
         }
 
-        // GET: Teacher/Create
+        
         [AuthorizeByRole("Admin")]
         public ActionResult Create()
         {
             return View();
         }
 
-        // POST: Teacher/Create
+        
         [HttpPost]
         [AuthorizeByRole("Admin")]
         [ValidateAntiForgeryToken]
@@ -38,7 +38,7 @@ namespace StudentManagerment2.Controllers
             return View(teacher);
         }
 
-        // GET: Teacher/Edit/5
+        
         [AuthorizeByRole("Admin")]
         public ActionResult Edit(int? id)
         {
@@ -46,15 +46,15 @@ namespace StudentManagerment2.Controllers
             {
                 return new HttpStatusCodeResult(System.Net.HttpStatusCode.BadRequest);
             }
-            Teacher teacher = db.Teachers.Find(id);
+            var teacher = db.Teachers.Find(id);
             if (teacher == null)
             {
                 return HttpNotFound();
             }
             return View(teacher);
         }
+
         
-        // POST: Teacher/Edit/5
         [HttpPost]
         [AuthorizeByRole("Admin")]
         [ValidateAntiForgeryToken]
@@ -69,7 +69,7 @@ namespace StudentManagerment2.Controllers
             return View(teacher);
         }
 
-        // GET: Teacher/Delete/5
+        
         [AuthorizeByRole("Admin")]
         public ActionResult Delete(int? id)
         {
@@ -77,7 +77,7 @@ namespace StudentManagerment2.Controllers
             {
                 return new HttpStatusCodeResult(System.Net.HttpStatusCode.BadRequest);
             }
-            Teacher teacher = db.Teachers.Find(id);
+            var teacher = db.Teachers.Find(id);
             if (teacher == null)
             {
                 return HttpNotFound();
@@ -85,15 +85,18 @@ namespace StudentManagerment2.Controllers
             return View(teacher);
         }
 
-        // POST: Teacher/Delete/5
+        
         [HttpPost, ActionName("Delete")]
         [AuthorizeByRole("Admin")]
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Teacher teacher = db.Teachers.Find(id);
-            db.Teachers.Remove(teacher);
-            db.SaveChanges();
+            var teacher = db.Teachers.Find(id);
+            if (teacher != null)
+            {
+                db.Teachers.Remove(teacher);
+                db.SaveChanges();
+            }
             return RedirectToAction("Index");
         }
     }
